requirements.txt
Django>=5.0
djangorestframework>=3.15
djangorestframework-simplejwt>=5.3
psycopg2-binary>=2.9   
Pillow>=10.0           
django-cors-headers>=4.3  
python-decouple>=3.8   
stripe>=9.0            
channels>=4.0          
channels-redis>=4.2    
gunicorn>=23.0
celery>=5.4
redis>=5.0
channels>=4.0
channels-redis>=4.2


